+++ lib/eggnstone_flutter.dart
@@ -11,4 +11,7 @@ export 'package:eggnstone_flutter/tools/DartTools.dart';
 export 'package:eggnstone_flutter/tools/LogTools.dart';
 export 'package:eggnstone_flutter/tools/PackageInfo.dart';
 export 'package:eggnstone_flutter/tools/UrlLauncherTools.dart';
+export 'package:eggnstone_flutter/tools/platform_tools/ClipboardNonWebTools.dart' if (dart.library.html) 'package:eggnstone_flutter/tools/platform_tools/ClipboardWebTools.dart';
+export 'package:eggnstone_flutter/tools/platform_tools/PlatformNonWebTools.dart' if (dart.library.html) 'package:eggnstone_flutter/tools/platform_tools/PlatformWebTools.dart';
 export 'package:eggnstone_flutter/widgets/BusyDialog.dart';
+export 'package:eggnstone_flutter/widgets/HyperLink.dart';