+++ lib/tools/platform_tools/ClipboardNonWebTools.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/services.dart';
+
+class ClipboardTools
+{
+    static void setData(String text)
+    => Clipboard.setData(ClipboardData(text: text));
+}