+++ lib/tools/platform_tools/PlatformNonWebTools.dart
@@ -1,0 +1,17 @@
+class PlatformTools
+{
+    static String get jsContextLocationHref
+    => '';
+
+    static set jsContextLocationHref(String value)
+    {}
+
+    static void htmlWindowHistoryPushState(String url)
+    {}
+
+    static dynamic htmlWindowDocumentGetElementById(String id)
+    => null;
+
+    static dynamic htmlWindowDocumentGetFirstElementByTagName(String tagName)
+    => null;
+}